Do not minify css code
Configures the USWDS gulp compiler to read CSS from assets/css (same as
it is already doing) but to ouput the minified CSS to dist/css. Then
updates Jekyll and the website to source CSS from dist/css and to ignore
the assets/css folder. Updates relevant .gitignore and .prettierignore
files to ignore the new dist folder and to format assets/css