Fix page titles
Removing the title attribute makes it so the pages don't have a title
which is bad for accessibility and for google analytics.

Also fixes a bug where the production jekyll build on deployment
populates the 'page.title' variable with the first <h1> on the page.
This behavior isn't replicated locally and caused a bug in the deployed
build.